% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r0density.R
\docType{package}
\name{r0density}
\alias{r0density}
\title{r0density}
\description{
r0density is a companion package for the paper "Population Density and Basic Reproductive Number of COVID-19 Revisited".
}
\details{
The package includes the data used in the analysis.
}
\author{
Antonio Paez, \email{paezha@mcmaster.ca}
}
