% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r0density.R
\docType{data}
\name{state_geo}
\alias{state_geo}
\title{States of the United States with population.}
\format{
A simple features data frame with 52 rows and 8 variables:
\describe{
  \item{GEO_ID}{Unique identifier of the state}
  \item{STATE}{Code of state}
  \item{NAME.x}{Short name of state}
  \item{LSAD}{Code of the Legal/Statistical Area Description}
  \item{GEOID}{Unique identifier of the county}
  \item{NAME.y}{Longer name of state}
  \item{population}{Total population from the 2010 Census (variable P001001)}
  \item{state}{Two letter code of state (from Sy et al.)}
  \item{geometry}{Geometry of boundaries}
}
}
\source{
US Bureau of Census
}
\usage{
data(state_geo)
}
\description{
Boundaries of the states in the United States.
}
\keyword{Census}
\keyword{States}
\keyword{United}
\keyword{covid-19}
\keyword{datasets}
