% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r0density.R
\docType{data}
\name{urban_geo}
\alias{urban_geo}
\title{Urban areas of the United States with population.}
\format{
A simple features data frame with 3,779 rows and 13 variables:
\describe{
  \item{UACE10}{2010 Census urban area code}
  \item{AFFGEOID10}{Long geographic identifier}
  \item{GEOID10}{Geographic identifier}
  \item{NAME10}{Short name of area}
  \item{LSAD10}{Code of the Legal/Statistical Area Description}
  \item{UATYP10}{2010 Census urban area type: C: Cluster with more than 2,500 but less than 50,000 population; U: Urbanized area with more than 50,000 population}
  \item{ALAND10}{2010 Census land area}
  \item{AWATER10}{2010 Census water area}
  \item{NAME}{Name of area}
  \item{population}{Total population from the 2010 Census (variable P001001)}
  \item{STATE}{Code of state}
  \item{state}{Two letter code of state (from Sy et al.)}
  \item{geometry}{Geometry of boundaries}
}
}
\source{
US Bureau of Census
}
\usage{
data(urban_geo)
}
\description{
Boundaries of urban areas in the United States.
}
\keyword{Census}
\keyword{States}
\keyword{United}
\keyword{covid-19}
\keyword{datasets}
